Fix terrible bug - I am a moron Add a task as a function, as previous approach was dumb. Working loop with motor stopped when tilt switch fires

+++ src/DHTSensor.cpp
@@ -25,7 +25,7 @@ const char* DHTSensor::status(){
 
 }
 
-void DHTSensor::init(const char *desc, int pin) {
+void DHTSensor::init(const char *desc, int8_t pin) {
   setPin(pin);
   setDesc(desc);
 }
